refactor(agentengine): split streaming_agent_run_with_events handler
